renamed library file `lib/.../metics.rb` to `lib/.../metrics.rb`
Basically looks like someone made a typo a while ago and rather than fixing the misspelling they used it so I replaced any refrences to it.
