+++ lib/sensu-plugins-mongodb/metrics.rb
